core: Blacklist 2021+ pixel experience XMLs from Google Photos
We want to include P21+ experience flags to enable new features,
however it seems that Google Photos uses it to decide whether to use the
TPU tflite delegate. There doesn't seem to be any fallback so we need to
make sure the feature is not exposed to the app so that a normal
NNAPI/GPU delegate can be used instead.

[ralph950412]:
However, PIXEL_2021_MIDYEAR_EXPERIENCE is included in Google Pixel 5a (barbet), meaning that it won't cause breakage.
Let's not blacklist it.

Test: Google Photos editor with PIXEL_{2021,2022,2023}_EXPERIENCE & PIXEL_{2022,2023}_MIDYEAR_EXPERIENCE system features in product
Change-Id: I349cffd7b15b6d9e91e2fb62d222fea52f69c40f
Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>
Signed-off-by: ralph950412 <ralph950412@gmail.com>