core: pm: Wipe package cache when upgrading system
As we keep adding and removing features with each build, it's necessary to have
package cache wiped out to avoid certain potential issues from arising due to a
dirty flash.

This commit is meant to supplement existing check for builds with incremental
version that starts with eng (not eng build type, which skips generating
package cache by design) which will continue to behave as expected.

Change-Id: Ic29f0291b4420c6cfdd91536cf6dbde2d1b91b56
Signed-off-by: Albert I <kras@raphielgang.org>